Wrap `fwrite` call in a retry loop
The `fwrite` function may write less bytes than in the given string, so
a loop is required to check the returned number of bytes written and
continue calling `fwrite` until the whole string has been written.

While we are at it, we can check for empty responses that possibly
indicate a write error. We can now set a configurable number of retries
with a configurable delay between each retry.