Fix missing increment on the index counter
While missing the increment operation, the post method never sent a
"Connection: Close" HTTP header. This resulted in a delay before the
connection is actually closed (maybe a timeout on the server side?).

This commit adds the missing increment operation and documents the
reason for using an index separate from the key variable in the
foreach loop.